Shows why we should really ignore ModifyAwareType for all Jackson beans
Pretty much only SOURCE and HASH (and NONE) modes work as we would expect. Relying on the ModifyAwareType is confusing or buggy.

So instead for ModifyAwareType beans lets just default to HASH based. Arguably we could do this for List, Set and Map as well. For all the Jackson ObjectMapper cases just have SOURCE, HASH or NONE (and ignore  ModifyAwareType).