Adds configurable destination package for querybeans
added compiler arguments to allow custom destination package suffix for querybeans.
motivation behind this is to allow Entities to be package protected.
With the .query + . assoc hardcoded this can't compile because the querybeans cannot find the enitities.
Kept the default by omission as it was (.query + .assoc).
Only did base querybean-generator, no changes were made to kotlin generator.